requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
django-restframework==0.0.1
django-webpack-loader==1.1.0
djangorestframework==3.12.4
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
